fix: proper conversion, better separation in domain layer
