+++ common/converter/api/src/commonMain/kotlin/presentation/decompose/ConverterComponent.kt
@@ -3,6 +3,8 @@ package presentation.decompose
 import com.arkivanov.decompose.router.slot.ChildSlot
 import com.arkivanov.decompose.value.Value
 import kotlinx.coroutines.flow.StateFlow
+import presentation.model.ConverterScreenState
+import presentation.model.TextFieldState
 
 interface ConverterComponent {
     val currencyListSlot: Value<ChildSlot<*, CurrencyListComponent>>
