+++ common/converter/api/src/commonMain/kotlin/presentation/model/CurrencyUiModel.kt
@@ -1,0 +1,9 @@
+package presentation.model
+
+data class CurrencyUiModel(
+    val currencyCode: String,
+    val currencyName: String,
+    val flagImageUrl: String,
+    val usdRate: Double,
+    val fetchDate: String
+)
