+++ common/converter/api/src/commonMain/kotlin/presentation/util/Mappers.kt
@@ -1,0 +1,20 @@
+package presentation.util
+
+import domain.model.CurrencyEntity
+import presentation.model.CurrencyUiModel
+import utils.toLocalDateTimeText
+
+fun CurrencyEntity.toUiModel(): CurrencyUiModel = CurrencyUiModel(
+    currencyCode = currencyCode,
+    currencyName = currencyName,
+    flagImageUrl = flagImageUrl,
+    usdRate = usdRate,
+    fetchDate = fetchTimeInstant.toLocalDateTimeText()
+)
+
+fun CurrencyUiModel.toEntity(): CurrencyEntity = CurrencyEntity(
+    currencyCode = currencyCode,
+    currencyName = currencyName,
+    flagImageUrl = flagImageUrl,
+    usdRate = usdRate
+)
